[PT-D][Tensor Parallel] Update the example for TP to use the new TP API and deprecate the old one (#1099)
* [PT-D][Tensor Parallel] Update the example for TP to use DTensor and new TP API
