--- distributed/sharded_tensor/requirements.txt
@@ -1,1 +1,0 @@
-torch>=1.12.0
