+++ distributed/tensor_parallelism/requirements.txt
@@ -1,0 +1,6 @@
+# Python dependencies required for running the example
+
+--pre
+--extra-index-url https://download.pytorch.org/whl/nightly/cu113
+--extra-index-url https://download.pytorch.org/whl/nightly/cu116
+torch >= 1.14.0.dev0; sys_platform == "linux"
