+++ run_python_examples.sh
@@ -60,9 +60,9 @@ function dcgan() {
 }
 
 function distributed() {
-    start
-    python sharded_tensor/tensor_parallel.py || error "tensor parallel example failed"
-    python ddp/main.py || error "ddp example failed" 
+    startrun
+    python distributed/tensor_parallelism/example.py || error "tensor parallel example failed"
+    python distributed/ddp/main.py || error "ddp example failed" 
 }
 
 function fast_neural_style() {
